cli/requirements.txt
click>=8.0.0
requests>=2.28.0
pyyaml>=6.0
tabulate>=0.9.0
colorama>=0.4.6
python-dateutil>=2.8.2